requirements.txt
protobuf
fastapi
pydantic
dotenv
python-dotenv
fastapi
uvicorn
gunicorn
python-multipart
google-generativeai